requirements.txt
hachoir==3.3.0
piexif==1.1.3
Pillow==11.1.0
